+++ src/piplexed/cli.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 import typer
 
 import piplexed
+from piplexed._print_table import print_list_table
+from piplexed._print_table import print_outdated_table
 from piplexed._print_tree import print_list_outdated
 from piplexed._print_tree import print_list_tree
 
@@ -23,11 +25,18 @@ def list(
         "-P",
         help="Include pre and dev releases in latest pypi version search",
     ),
+    table: bool = typer.Option(False, "--table", "-T", help="print output as a table"),
 ) -> None:
-    if outdated and is_prelease:
+    if outdated and is_prelease and table:
+        print_outdated_table(piplexed.find_outdated_packages(stable=False))
+    elif outdated and is_prelease:
         print_list_outdated(piplexed.find_outdated_packages(stable=False))
+    elif outdated and table and not is_prelease:
+        print_outdated_table(piplexed.find_outdated_packages(stable=True))
     elif outdated and not is_prelease:
         print_list_outdated(piplexed.find_outdated_packages(stable=True))
+    elif table:
+        print_list_table(piplexed.get_pipx_metadata())
     else:
         print_list_tree(piplexed.get_pipx_metadata())